requirements.txt
fastapi==0.128.8
uvicorn[standard]==0.40.0
openai==2.21.0
python-dotenv==1.2.1